consistency - set all long fields to ints to match the data on the backend backend

+++ src/org/mythtv/db/dvr/ProgramGuideDaoHelper.java
@@ -107,7 +107,7 @@ public class ProgramGuideDaoHelper extends ProgramDaoHelper {
 	 * @see org.mythtv.db.dvr.ProgramDaoHelper#findOne(int, org.joda.time.DateTime)
 	 */
 	@Override
-	public Program findOne( Long channelId, DateTime startTime ) {
+	public Program findOne( int channelId, DateTime startTime ) {
 		Log.d( TAG, "findOne : enter" );
 		
 		String selection = ProgramConstants.TABLE_NAME_PROGRAM + "." + ProgramConstants.FIELD_CHANNEL_ID + " = ? AND " + ProgramConstants.TABLE_NAME_PROGRAM + "." + ProgramConstants.FIELD_START_TIME + " = ?";
